FIX(Group&Job): 对规范性进行调整
FEAT(BaseController): 增加评判结果的多态，可传入自定义成功失败消息
REFACTOR(PROJECT): swagger文档按分组进行展示，各分包结构规范性名称变更，位置移动，根据新方法进行变动

+++ src/main/java/com/breze/service/rbac/impl/JobServiceImpl.java
@@ -2,7 +2,11 @@ package com.breze.service.rbac.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.breze.common.enums.ErrorEnum;
+import com.breze.common.exception.BusinessException;
+import com.breze.entity.pojo.rbac.GroupJob;
 import com.breze.entity.pojo.rbac.Job;
+import com.breze.mapper.rbac.GroupJobMapper;
 import com.breze.mapper.rbac.JobMapper;
 import com.breze.service.rbac.JobService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,6 +27,9 @@ public class JobServiceImpl extends ServiceImpl<JobMapper, Job> implements JobSe
     @Autowired
     public JobMapper jobMapper;
 
+    @Autowired
+    public GroupJobMapper groupJobMapper;
+
     @Override
     public List<Job> findAll(String name) {
         return jobMapper.selectList(new QueryWrapper<Job>().like("name", name));
@@ -37,9 +44,9 @@ public class JobServiceImpl extends ServiceImpl<JobMapper, Job> implements JobSe
     public List<Job> searchOr(Job job) {
         QueryWrapper<Job> qw = new QueryWrapper<>();
         qw.like("id", job.getId()).or()
-            .like("name", job.getName()).or()
-            .like("state", job.getState()).or()
-            .like("remark", job.getRemark());
+                .like("name", job.getName()).or()
+                .like("state", job.getState()).or()
+                .like("remark", job.getRemark());
         return jobMapper.selectList(qw);
     }
 
@@ -66,7 +73,7 @@ public class JobServiceImpl extends ServiceImpl<JobMapper, Job> implements JobSe
     @Transactional(rollbackFor = Exception.class)
     public Boolean update(Job job) {
         Job j = jobMapper.selectOne(new QueryWrapper<Job>().eq("name", job.getName()));
-        if (j.getName().equals(job.getName()) ) {
+        if (j.getName().equals(job.getName())) {
             if (j.getId().equals(job.getId())) {
                 return jobMapper.updateById(job) > 0;
             }
@@ -78,6 +85,15 @@ public class JobServiceImpl extends ServiceImpl<JobMapper, Job> implements JobSe
     @Override
     @Transactional(rollbackFor = Exception.class)
     public Boolean delete(Long id) {
-        return jobMapper.deleteById(id) > 0;
+        try {
+            if (groupJobMapper.selectCount(new QueryWrapper<GroupJob>().eq("job_id", id)) > 0) {
+                throw new BusinessException(ErrorEnum.FindException, "删除岗位失败");
+            }
+            return jobMapper.deleteById(id) > 0;
+        } catch (Exception e) {
+            throw new BusinessException(ErrorEnum.FindException, "删除岗位失败");
+        }
+
+
     }
 }
